Struct curs::MultipartBodyBuilder

pub struct MultipartBodyBuilder {
    // some fields omitted
}

You're not expected to be using the MultipartBodyBuilder on your own, Make a Request and it will know when to delegate if there are any files to be posted. It's still exported publicly because it may come in handy for other uses.

Methods

impl MultipartBodyBuilder

fn new() -> MultipartBodyBuilder

fn build<'a>(self, files: Vec<FileUpload>, params: Params<'a>) -> Result<MultipartBodyBuilder, CursError>